Move BuildExecutionStatus to constants module
