+++ pnc-mock/src/main/java/org/jboss/pnc/mock/executor/BuildExecutorStatusChangedEventMock.java
@@ -18,7 +18,7 @@
 
 package org.jboss.pnc.mock.executor;
 
-import org.jboss.pnc.spi.BuildExecutionStatus;
+import org.jboss.pnc.enums.BuildExecutionStatus;
 import org.jboss.pnc.spi.BuildResult;
 import org.jboss.pnc.spi.events.BuildExecutionStatusChangedEvent;
