+++ spi/src/main/java/org/jboss/pnc/spi/executor/BuildExecutionSession.java
@@ -18,7 +18,7 @@
 
 package org.jboss.pnc.spi.executor;
 
-import org.jboss.pnc.spi.BuildExecutionStatus;
+import org.jboss.pnc.enums.BuildExecutionStatus;
 import org.jboss.pnc.spi.builddriver.BuildDriverResult;
 import org.jboss.pnc.spi.environment.RunningEnvironment;
 import org.jboss.pnc.spi.executor.exceptions.ExecutorException;
